feat: fix some indetation

+++ src/Controller/FacturationModelController.php
@@ -1,17 +1,26 @@
 <?php
 
 namespace App\Controller;
+use App\Entity\FacturationModel;
 use App\Repository\FacturationModelRepository;
+use App\Repository\ClientRepository;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use App\Entity\FacturationModel;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Contracts\Cache\TagAwareCacheInterface;
 use Symfony\Bundle\SecurityBundle\Security;
+use App\Controller\Account;
 
 #[Route('/api/facturation-model')]
-
+#[IsGranted("ROLE_ADMIN", message: "Hanhanhaaaaan vous n'avez pas dit le mot magiiiiqueeuuuuuh")]
 class FacturationModelController extends AbstractController
 {
     private $user;
@@ -77,7 +86,6 @@ class FacturationModelController extends AbstractController
 
         $data = $request->toArray();
         if (isset($data['client'])) {
-
             $client = $clientRepository->find($data["client"]);
         }
 
@@ -106,9 +114,10 @@ class FacturationModelController extends AbstractController
 
         $data = $request->toArray();
         if (isset($data['force']) && $data['force'] === true) {
+            if (!$this->isGranted("ROLE_ADMIN")) {
+                return new JsonResponse(["error" => "Hanhanhaaaaan vous n'avez pas dit le mot magiiiiqueeuuuuuh"], JsonResponse::HTTP_FORBIDDEN);
+            }
             $entityManager->remove($facturationModel);
-
-
         } else {
             $facturationModel
                 ->setStatus("off")
@@ -117,9 +126,6 @@ class FacturationModelController extends AbstractController
 
             $entityManager->persist($facturationModel);
         }
-
-
-
         $entityManager->flush();
         $cache->invalidateTags(["facturationModel"]);
         return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
